requirements.txt
flask==2.3.3
flask-cors==3.0.10
werkzeug==2.3.7
numpy>=1.24.3
opencv-python>=4.7.0
pillow>=9.5.0
ultralytics>=8.0.196
torch>=2.0.0
matplotlib>=3.7.1
gunicorn
python-dotenv==1.0.0